Added some lint warnings, cleaned up some code

+++ src/algorithms/population.rs
@@ -63,8 +63,8 @@ pub(crate) fn generate_populations<Progress: ProgressObserver>(target: &mut Worl
                         suitability += 15.0 // estuaries are liked
                     }
                     if let Some(water_cell) = tile.harbor_tile_id {
-                        let water_cell = tiles.try_entity_by_id::<TileForPopulationNeighbor>(&(water_cell as u64))?;
-                        if let Some(lake_type) = &water_cell.lake_id.map(|id| lake_map.try_get(&(id as u64))).transpose()?.map(|l| &l.type_) {
+                        let water_cell = tiles.try_entity_by_id::<TileForPopulationNeighbor>(&water_cell)?;
+                        if let Some(lake_type) = &water_cell.lake_id.map(|id| lake_map.try_get(&id)).transpose()?.map(|l| &l.type_) {
                             match lake_type {
                                 LakeType::Fresh => suitability += 30.0,
                                 LakeType::Salt => suitability += 10.0,
